fix: rename css files from index to styles
