+++ app/components/templates/base-template/index.tsx
@@ -1,5 +1,5 @@
 // https://nextjs.org/docs/basic-features/built-in-css-support
-import styles from './index.module.sass';
+import styles from './styles.module.sass';
 import { IBaseTemplate } from './types';
 
 const BaseTemplate: React.FC<IBaseTemplate> = ({ sampleTextProp }) => {
